requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
psycopg2-binary==2.7.6.1
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
six==1.12.0
SQLAlchemy==1.3.5
Werkzeug==0.15.4
